python-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.6.3
python-dotenv==1.0.1
edge-tts==6.1.10
pypdf==4.1.0
openai-whisper==20231117
torch==2.2.1
torchaudio==2.2.1
librosa==0.10.1
numpy==1.26.4
numba==0.59.0
resampy==0.4.3
soundfile==0.12.1
gunicorn==21.2.0
